Added support for CRS to csv point data
The changes allow a Coordinate Reference System (CRS) name to be passed as an option to be used when converting a csv2geojson.  The CRS name is then included when creating the new GeoJSON file which may not be using lat/long data, but instead might be using a custom x/y coordinate system, or alternative spatial model. 
This allows this leaflet plug-in to work with Proj4-leaflet in projecting custom coordinate csv data onto a map.